# rpc/grpc/proto/types.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "types.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/chain/version": {
      "get": {
        "operationId": "Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoVersionResponse"
            }
          }
        },
        "tags": [
          "ChainAPI"
        ]
      }
    },
    "/orchestra/exec/create": {
      "post": {
        "operationId": "ExecCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonRequest"
            }
          }
        ],
        "tags": [
          "OrchestraAPI"
        ]
      }
    },
    "/orchestra/exec/find": {
      "post": {
        "operationId": "ExecFind",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonRequest"
            }
          }
        ],
        "tags": [
          "OrchestraAPI"
        ]
      }
    },
    "/orchestra/exec/get": {
      "post": {
        "operationId": "ExecGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoOrchestraCommonRequest"
            }
          }
        ],
        "tags": [
          "OrchestraAPI"
        ]
      }
    },
    "/order/change": {
      "post": {
        "operationId": "ChangeOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrderId"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoChangeOrderParam"
            }
          }
        ],
        "tags": [
          "OrderAPI"
        ]
      }
    },
    "/order/create": {
      "post": {
        "operationId": "CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrderId"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoCreateOrderParam"
            }
          }
        ],
        "tags": [
          "OrderAPI"
        ]
      }
    },
    "/order/info": {
      "get": {
        "operationId": "GetOrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrderInfo"
            }
          }
        },
        "parameters": [
          {
            "name": "internalId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OrderAPI"
        ]
      }
    },
    "/order/terminate": {
      "post": {
        "operationId": "TerminateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoOrderId"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoTerminateOrderParam"
            }
          }
        ],
        "tags": [
          "OrderAPI"
        ]
      }
    }
  },
  "definitions": {
    "protoChangeConnectionParam": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string"
        },
        "dynamicParam": {
          "$ref": "#/definitions/protoConnectionDynamicParam"
        }
      }
    },
    "protoChangeOrderParam": {
      "type": "object",
      "properties": {
        "privacy": {
          "$ref": "#/definitions/protoContractPrivacyParam"
        },
        "buyer": {
          "$ref": "#/definitions/protoUser"
        },
        "seller": {
          "$ref": "#/definitions/protoUser"
        },
        "changeConnectionParam": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoChangeConnectionParam"
          }
        }
      }
    },
    "protoConnectionDynamicParam": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "quoteId": {
          "type": "string"
        },
        "quoteItemId": {
          "type": "string"
        },
        "connectionName": {
          "type": "string"
        },
        "paymentType": {
          "type": "string"
        },
        "billingType": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "serviceClass": {
          "type": "string"
        },
        "bandwidth": {
          "type": "string"
        },
        "billingUnit": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        },
        "Addition": {
          "type": "number",
          "format": "float"
        },
        "startTime": {
          "type": "string",
          "format": "int64"
        },
        "startTimeStr": {
          "type": "string"
        },
        "endTime": {
          "type": "string",
          "format": "int64"
        },
        "endTimeStrTimeStr": {
          "type": "string"
        }
      }
    },
    "protoConnectionParam": {
      "type": "object",
      "properties": {
        "staticParam": {
          "$ref": "#/definitions/protoConnectionStaticParam"
        },
        "dynamicParam": {
          "$ref": "#/definitions/protoConnectionDynamicParam"
        }
      }
    },
    "protoConnectionStaticParam": {
      "type": "object",
      "properties": {
        "itemId": {
          "type": "string"
        },
        "buyerProductId": {
          "type": "string"
        },
        "ProductOfferingId": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "srcCompanyName": {
          "type": "string"
        },
        "srcRegion": {
          "type": "string"
        },
        "srcCity": {
          "type": "string"
        },
        "srcDataCenter": {
          "type": "string"
        },
        "srcPort": {
          "type": "string"
        },
        "dstCompanyName": {
          "type": "string"
        },
        "dstRegion": {
          "type": "string"
        },
        "dstCity": {
          "type": "string"
        },
        "dstDataCenter": {
          "type": "string"
        },
        "dstPort": {
          "type": "string"
        }
      }
    },
    "protoContractPrivacyParam": {
      "type": "object",
      "properties": {
        "privateFrom": {
          "type": "string"
        },
        "privateFor": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "privateGroupID": {
          "type": "string"
        }
      }
    },
    "protoCreateOrderParam": {
      "type": "object",
      "properties": {
        "privacy": {
          "$ref": "#/definitions/protoContractPrivacyParam"
        },
        "buyer": {
          "$ref": "#/definitions/protoUser"
        },
        "seller": {
          "$ref": "#/definitions/protoUser"
        },
        "connectionParam": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoConnectionParam"
          }
        }
      }
    },
    "protoOrchestraCommonRequest": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoOrchestraCommonResponse": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        },
        "resultCount": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoOrderId": {
      "type": "object",
      "properties": {
        "internalId": {
          "type": "string"
        }
      }
    },
    "protoOrderInfo": {
      "type": "object",
      "properties": {
        "buyer": {
          "$ref": "#/definitions/protoUser"
        },
        "seller": {
          "$ref": "#/definitions/protoUser"
        },
        "orderId": {
          "type": "string"
        },
        "orderType": {
          "type": "string"
        },
        "orderState": {
          "type": "string"
        },
        "contractState": {
          "type": "string"
        },
        "connections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoConnectionParam"
          }
        },
        "track": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoOrderLifeTrack"
          }
        }
      }
    },
    "protoOrderLifeTrack": {
      "type": "object",
      "properties": {
        "contractState": {
          "type": "string"
        },
        "orderState": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "time": {
          "type": "string",
          "format": "int64"
        },
        "Hash": {
          "type": "string"
        }
      }
    },
    "protoTerminateConnectionParam": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string"
        },
        "dynamicParam": {
          "$ref": "#/definitions/protoConnectionDynamicParam"
        }
      }
    },
    "protoTerminateOrderParam": {
      "type": "object",
      "properties": {
        "privacy": {
          "$ref": "#/definitions/protoContractPrivacyParam"
        },
        "buyer": {
          "$ref": "#/definitions/protoUser"
        },
        "seller": {
          "$ref": "#/definitions/protoUser"
        },
        "terminateConnectionParam": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoTerminateConnectionParam"
          }
        }
      }
    },
    "protoUser": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "protoVersionResponse": {
      "type": "object",
      "properties": {
        "buildTime": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "hash": {
          "type": "string"
        }
      }
    }
  }
}